refactored CompositeCurveType parsing to work for gml 3 as well
